Remove TabPanel class
Changes:
- Move TabPanel logic to EntityTabPanel
- Delete TabPanel class
- Updated EntityContentPanel to use EntityTabPanel<?> instead of TabPanel
- DockingWindow changes minumum size of Entity TabPanel to avoid inital scrolling

Ref #1798
